workers/requirements.txt
google-cloud-pubsub==2.18.4
google-cloud-bigquery==3.11.4

